% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data109det}
\alias{data109det}
\title{Data from experiment ALEXIS 109: Detection experiment}
\format{A data frame with 60,480 rows and 13 columns.
\describe{
  \item{id}{Participant unique id number}
  \item{date}{Date of testing}
  \item{time}{Time of day for each trial}
  \item{day}{Rank order of date the data was collected}
  \item{session}{Session number, there were 13 sessions, run on the same day}
  \item{trial}{Trial number within session, from 1 to 60}
  \item{ici}{Inter-click interval in ms, Note: -1 is for sessions with no lead click}
  \item{ild}{Inter-arual-level difference of variable, 0 or 10 dB}
  \item{itd}{Inter-arual-time difference of variable, 0 or 0.35 ms}
  \item{ratio}{dB-difference between lead and lag click of variable}
  \item{revno}{Reversal number, equals 0 if the trial was not a reversal}
  \item{correct}{Lag click present (1) or absent (0)}
  \item{answer}{The answer given by the listener, "left" (1) or "right" (0).
                The answer was correct if answer == correct}
  }}
\usage{
data109det
}
\description{
This is data from an experiment with three participants, tested on detection
of a lag click following a lead click after an inter-click interval between 
0.125 to 256 ms. The binaural difference in the lag-click was an ITD of 0.35 ms
an ILD of 10 dB or both or neither.
}
\keyword{datasets}
