% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ths_lat}
\alias{ths_lat}
\title{Processed data from experiment ALEXIS 109: Lateralization experiment}
\format{A data frame with 117 rows and 11 columns.
\describe{
  \item{id}{Participant unique id number}
  \item{ici}{Inter-click interval in ms, Note: -1 is for sessions with no lead click}
  \item{bincond}{Binaural condition: 1 = ILD-only, 2 = ITD-only, 3 = ITD+ILD}
  \item{thmode}{Threshold estimate (mode of posterior distribution), i.e., LLR [dB] at 75 percent correct responses}
  \item{X.0.95}{Lower 0.95 credibility interval of posterior distribution of threshold}
  \item{X0.95.}{Upper 0.95 credibility interval of posterior distribution of threshold}
  \item{mmode}{Mode of location parameter m}
  \item{wmode}{Mode of scale parameter w (spread)}
  \item{lapsemode}{Mode of lapse-reate parameter lambda}
  \item{mineff}{MCMC diagnostic: lowest (worst) effective sample size over estimated parameters}
  \item{maxrhat}{MCMC diagnostic: highest (worst) rhat (should be very close to 1.00)}
  }}
\usage{
ths_lat
}
\description{
This is processed data (estimated thresholds) from an experiment with three participants, tested on lateralization
of a large binaural difference in lag clicks following a lead click after
an inter-click interval between 0.125 to 256 ms. The binaural difference was
an ITD of 0.35 ms or an ILD of 10 dB or both.
}
\keyword{datasets}
